Cleans up simplye/shared gradle to pull in .so files.
Removes the extra commented code that's not used in the
simplified-app-simplye/build.gradle, and includes the native_dependencies
to the build.gradle for simplified-app-shared. This is pulling in the .so
file from the nexus repository, but still not finding it when we try to
invoke the ScanSettings class at runtime.