Be more careful about book content types
Prior to this change, if a server served a book as
application/octet-stream, the application would reject it. This set
of changes improves content handling so that:

  * If the feed promises content of type A, and the server delivers
    content of type B, and A == B, then the content is accepted.

  * If the feed promises content of type A, and the server delivers
    content of type application/octet-stream, then the content
    is accepted.

  * Otherwise, the content is rejected. The exception raised will
    be of type BookUnexpectedTypeException.

In all cases, if A == application/octet-stream, the book
database will reject the content after the above rules have
been evaluated. This will actually raise the same exception as
seen in https://jira.nypl.org/browse/SIMPLY-1388, it's just that
this exception can now only be encountered if the feed promises
application/octet-stream, which is very unlikely and would be a
mistake anyway.

Fix: https://jira.nypl.org/browse/SIMPLY-1388